Const all the ask functions.
Here we have removed memoization on some functions that use tags for
memoization. A few tests are failing because there is no longer this
implicit creation of tags happening. The removal of creating the tags
is a breaking change and may be hard to get people on board with.

But, the code is much nicer now that const correctness has started to
propagate through.

Next steps:
  1) add memoization back in using cache data in mesh class rather than
     tags
  2) factor out the mutuable cache data out of the mesh class
  3) update tests